readmes and cowmangliacoms
updated the readmes to be .md, except for the one inside the cowmangliacoms port, as its original filetype was txt. i've also fixed a few things about the port that were added by me that were NOT in the original cowmangliacoms version, such as a null movement script. i don't even know how that got in there, probably just merged my personal config with it on accident.